PBP: The PBP extension needs to be upper-case.
EBOOT.PBP is an archive format used for PS1 Classics games on PSP and PS3
(and PS4 and probably also PS5)
On PSP case does not matter since the file will be stored in a FAT
filesystem but on PS3 it does matter as it will be stored in a UFS2
filesystem which is case sensitive.

EBOOT.PBP files from the official PS Store comes wrapped in a PKG container
and installes as EBOOT.PBP uppercase. Thus other tools that create
EBOOT.PBP files also create them in uppercase such as POP-FE and PSX2PSP.

Change the *.pbp file selector filter to *.PBP so that the file selector
will find these files on platforms with case-sensitive filesystems.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>