Add "Advanced Settings" option for rounding texture coordinates from an upscaled native resolution
As per the title; normal DuckStation behavior is to floor the texture coordinates only when the native resolution is upscaled, instead of rounding them as done when the native resolution is unchanged (due to very blatant errors in some games). This adds a setting to force texture coordinates to be rounded even when the game is upscaled, as this provides a more accurate (relative to what the game looks like in its original resolution) behavior in certain games, such as the Crash Bandicoot series.