GameDB: Set DisablePGXPOn2DPolygons on Xenogears
Fixes misaligned UI. For future reference: game does a multiply by 4096,
then a right shift of 12, which should give back the same value, but
with it being implemented in PGXP-CPU with a divide that does not round
down, it ends up being off by one.

Rounding down loses precision in other games, so that's not an option.