Qt: Drop -settings/-portable CLI parameters
This hasn't been tested in years, and doesn't makes sense when you
consider other files are saved in the DataRoot directory (e.g. custom
properties, play time, crash dumps, etc).

-portable is redundant as the presence of settings.ini in the AppRoot
will trigger it anyway, therefore it is only relevant for first-time
installs. In which case it's just as much work to create an empty
settings.ini or portable.txt file as it is to create a shortcut with
-portable on the command line.

It was also broken anyway for a long time, and nobody complained.