Scripts: Remove PKGBUILD
I originally provided this an alternative to the broken AUR packages.

However, it seems that Arch users would rather use broken packages and
keep complaining to me instead of their packager. I specifically forbid
packages for DuckStation (see README.md), and there's no way to request
removal of these packages without handing my details over to a
distribution I want nothing to do with.

So this is step one. Next step will be removing Linux support entirely,
because I'm sick of the headaches and hacks for an operating system that
only compromises 2% of the userbase, and I don't even use myself. But I'm
hoping the Linux community will be reasonable, because as someone giving
up my free time and not being compensated in any way, I shouldn't have
to deal with this.

Just grep the source for "wayland" and you'll see what I mean.