CI: Remove Flatpak builds
Now they're getting random timeouts updating the SDKs, just tired of it.

The workflow to publish is still in the tree, but I haven't updated
Flathub in months, and it's arguably kinda pointless since the Flatpak
is literally just a wrapper of the AppImage now...

So, still undecided what to do about that. If you are using it, please
let me know, because in the months that it's been deprecated only a
couple of people mentioned that they were using it. Therefore, as best I
can tell there's basically no userbase.