Scripts: Remove Flatpak manifest
As I noted in 41b8f2bf77962cdbb239f7c02bbd15dff3e77f17 and #3454, I was
still undecided on how to proceed with the Flatpak going forward, and
was waiting on user feedback.

However, it appears that decision has been made for me. As of the time,
of writing, the page displays "This application is no longer available on
Flathub." "This application is no longer maintained on Flathub", and I
was not notified in any way of the removal.

No idea if this was automated (it's been ~3 months since the last push),
or manual, so I'm not pointing fingers at anyone here. But it is still
completely unprofessional on flathub's part. No other distributor would
remove applications without notifying the developer. Combined with the
constant workflow/CI breakage, new requirements being added and enforced
with little notice, and delisting of anything more than a couple years
old, I would recommend to any application developers considering
flatpak/hub and don't heavily use it themselves: don't, it ain't worth it.

No point keeping it around now, since it was only a copy of the AppImage
anyway, and if I recall correctly there's other tools that can bwrap an
AppImage if you really desire. But IMO the security argument is kinda
weak, if you don't trust the code you're running, don't run it. The app
is still interacting with the rest of the OS regardless...