GPU/HW: Draw small fill sprites with software if TC is enabled
Wild Arms 2 clears out the sprite area of VRAM, then uploads new
sprites. The clears were not counting as fills before, because
they did not meet the page threshold.

It's not worth risking treating all untextured sprite draws as
clears, so instead just draw it with software as well to ensure
the local copy of VRAM is cleared.