README: Remove non-functional link
Been paying hundreds of dollars for hosting for the last few years, and
with no immediate plans to push a new Android release, it does not make
sense for me to keep paying for it, especially since I make $0 off
DuckStation.

(it's been non-functional for a few weeks, apparently it took until now
for someone else to notice, which kinda shows how little use it was
getting)

If there's significant demand, then I may re-consider.